Simplify Postgres store

+++ crates/kagzi-server/src/scheduler.rs
@@ -105,14 +105,17 @@ pub async fn run(store: PgStore, settings: SchedulerSettings, shutdown: Cancella
         }
         let now = Utc::now();
 
-        let due: Vec<WorkflowSchedule> =
-            match store.schedules().due_schedules(now, batch_size).await {
-                Ok(due) => due,
-                Err(e) => {
-                    error!("Failed to fetch due schedules: {:?}", e);
-                    continue;
-                }
-            };
+        let due: Vec<WorkflowSchedule> = match store
+            .schedules()
+            .due_schedules(now, batch_size as i64)
+            .await
+        {
+            Ok(due) => due,
+            Err(e) => {
+                error!("Failed to fetch due schedules: {:?}", e);
+                continue;
+            }
+        };
 
         if due.is_empty() {
             continue;
@@ -136,8 +139,6 @@ pub async fn run(store: PgStore, settings: SchedulerSettings, shutdown: Cancella
             }
 
             let remaining_budget = (max_workflows_per_tick - created_this_tick) as usize;
-            // Use `last_fired_at` if available; otherwise, shift `next_fire_at` back by 1ns
-            // so that `schedule.after(&from)` includes the scheduled fire time itself.
             let from = schedule
                 .last_fired_at
                 .unwrap_or_else(|| schedule.next_fire_at - chrono::Duration::nanoseconds(1));
